% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbp.R
\docType{data}
\name{cbp}
\alias{cbp}
\title{A colorblind-friendly palette}
\format{the length of a vector, eight.}
\usage{
data(cbp)
}
\description{
The character vector of colorblind-friendly palette
}
\examples{
# Confirm color code.
data(cbp)
barplot(setNames(rep(1,8), cbp),
border = cbp, col=cbp, yaxt="n", las=2)
}
\keyword{datasets}
