% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cge.R
\docType{data}
\name{cge}
\alias{cge}
\title{KEGG Correspondence table of Chinese hamster.}
\format{An object of data frame with 25207 rows and 10 columns(at 11/16/17), named "GENE", "Description", "ecs", "GeneID", "RefIDs",
    "Symbol", "Ref.protein", "UPIDs", "KOs", "MAPs".}
\usage{
data(cge)
}
\description{
The data from the KEGG data base of Chinese hamster(09/15/17).
   This data colected from KEGG using KEGG API and NCBI gene(gene2refseq.gz).
}
\examples{
data(cge)
cge[1:3, c("GENE","Description")]

}
\references{
\href{https://rpubs.com/nishikosh/297854}{Rpubs}
}
\keyword{datasets}
