% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cge100.R
\docType{data}
\name{cge100}
\alias{cge100}
\title{cge100}
\format{An object of data.frame consists from the rows as 100 genes and the columns as 14 days.}
\usage{
data(cge100)
}
\description{
The data.frame of pathview input with organism code.
}
\examples{
head(cge100)
}
\keyword{datasets}
