% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdat.R
\docType{data}
\name{cpdat}
\alias{cpdat}
\title{cpdat}
\format{An object of dataframe with with 40 rows and 2 columns with rownames of KEGG compound IDs.}
\usage{
data(cpdat)
}
\description{
Metabolome data for KEGG mapping using pathview. Row names must be KEGG compound IDs.
}
\examples{
data(cpdat)
head(cpdat)
}
\keyword{datasets}
