% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat1.R
\docType{data}
\name{dat1}
\alias{dat1}
\title{dat1}
\format{An object of dataframe with with 18 rows and 14 columns.}
\usage{
data(dat1)
}
\description{
multivariate data.
}
\examples{
data(dat1)
head(dat1)
}
\keyword{datasets}
