% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2.R
\docType{data}
\name{dat2}
\alias{dat2}
\title{dat2}
\format{An object of dataframe with with 18 rows and 5 columns.}
\usage{
data(dat2)
}
\description{
Multivariate compression value with three factor attributes.
}
\examples{
data(dat2)
head(dat2)
}
\keyword{datasets}
