% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2norep.R
\docType{data}
\name{dat2norep}
\alias{dat2norep}
\title{dat2norep}
\format{An object of dataframe with with 6 rows and 5 columns.}
\usage{
data(dat2norep)
}
\description{
Multivariate compression value with two factor attributes, excluding repetition factor of dat2
}
\examples{
data(dat2norep)
head(dat2norep)
\dontrun{
## create 'dat2norep' from 'dat2'
dat2 \%>\%
 group_by(strains, days, reps) \%>\%
 summarise_at(vars(v), list(~mean)) \%>\%
 ungroup() \%>\%
 tibble::add_column(name=c("A.d1", "A.d2", "A.d3", "B.d1", "B.d2", "B.d3"),
     .before = "strains")
}

}
\keyword{datasets}
