% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpkm.R
\docType{data}
\name{fpkm}
\alias{fpkm}
\title{fpkm}
\format{An object of dataframe with 1000 rows and 12 columns.}
\usage{
data(fpkm)
}
\description{
The FPKM dataframe from RNA-seq.
}
\examples{
data(fpkm)
fpkm[1:6,1:6]
}
\keyword{datasets}
