% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegorg.R
\docType{data}
\name{kegorg}
\alias{kegorg}
\title{KEGG organism data coverted to data.frame.}
\format{An object of dataframe with with 3 columns.}
\usage{
data(kegorg)
}
\description{
The data from the KEGG data base of KEGG ORGANISM(11/16/17). Those of  entries  which, mapped on PATHWAY map.
   This data colected from KEGG using KEGG API.
}
\examples{
data(kegorg)
kegorg[1:3,]

}
\references{
\href{https://rpubs.com/nishikosh/297854}{Rpubs}
}
\keyword{datasets}
