% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kodat.R
\docType{data}
\name{kodat}
\alias{kodat}
\title{kodat}
\format{An object of dataframe with 300 rows and 2 columns with rownames of KEGG ortholog IDs.}
\usage{
data(kodat)
}
\description{
Transcriptome data for KEGG mapping using pathview. Row names must be gene IDs, KEGG ortholog IDs.
}
\examples{
data(kodat)
head(kodat)
}
\keyword{datasets}
