% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_iupac.R
\docType{data}
\name{n_iupac}
\alias{n_iupac}
\title{IUPAC nucleotide code}
\format{the data.frame}
\usage{
data(n_iupac)
}
\description{
The character vector of IUPAC nucleotide code
}
\examples{
# IUPAC nucleotide code.
data(n_iupac)
}
\keyword{datasets}
