% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfpkm.R
\docType{data}
\name{nfpkm}
\alias{nfpkm}
\title{nfpkm}
\format{An object of dataframe with 12 rows and 1003 columns.}
\usage{
data(fpkm)
}
\description{
The normalized FPKM dataframe from RNA-seq, using TMM normalization. The first four columns is the categolical information of 12 samples.
   The rest of 1000 columns is FPKM of genes.
}
\examples{
data(nfpkm)
nfpkm[1:6,1:6]
}
\keyword{datasets}
