% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_ascii.R
\docType{data}
\name{q_ascii}
\alias{q_ascii}
\title{Quality Score Encoding}
\format{An object of data frame with 41 rows and 3 colmns.}
\usage{
data(q_ascii)
}
\description{
the quality score is represented as the character with an ASCII code equal to its value + 33.
 The following table demonstrates the relationship between the encoding character, its ASCII code, and the quality score represented.
}
\examples{
data(q_ascii)
}
\keyword{datasets}
