% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_tab.R
\docType{data}
\name{re_tab}
\alias{re_tab}
\title{The Recognition Site of Restriction Enzymes}
\format{the data.frame}
\usage{
data(re_tab)
}
\description{
The Recognition Site of Restriction Enzymes
}
\examples{
# IUPAC nucleotide code.
data(re_tab)
}
\references{
\href{http://catalog.takara-bio.co.jp/product/basic_info.php?unitid=U100003632}{takara_bio}
}
\keyword{datasets}
