% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_equas_ko.R
\docType{data}
\name{react_equas_ko}
\alias{react_equas_ko}
\title{The Equations from KEGG REACTION database coverted to dataframe.}
\format{An object of data frame with 28326 rows and 5 columns(at 03/03/19), named "Rid", "n",
 "Cid", "equation", "Kids".}
\usage{
data(react_equas_ko)
}
\description{
The data from the KEGG data base of KEGG REACTION(03/03/19). Those of 11137 entries  which, mapped on PATHWAY map.
 This data colected from KEGG REACTION database using KEGG API.
 linked by K number
}
\examples{
\dontrun{
data(react_equas_ko)
head(react_equas_ko)
}
}
\references{
\href{https://rpubs.com/nishikosh/kegg}{Rpubs}
}
\keyword{datasets}
