% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.degv.R
\docType{data}
\name{res.degv}
\alias{res.degv}
\title{res.degv}
\format{An object of data.frame consists from kegg orthology ids and compound ids per pathways.}
\usage{
data(res.degv)
}
\description{
The result data.frame of rsko::degview.
}
\examples{
head(res.degv)
}
\keyword{datasets}
