% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\docType{data}
\name{vcf}
\alias{vcf}
\title{Variant Call Format}
\format{An object of data frame with 1000 rows and 15 columns.}
\usage{
data(vcf)
}
\description{
The data.frame of VCF format without header, contains six sampled vcf.
}
\examples{
\dontrun{
data(vcf)
head(vcf)
}
}
\references{
\href{https://samtools.github.io/hts-specs/VCFv4.2.pdf}{VCFv4.2}
}
\keyword{datasets}
